requirements.txt
streamlit
python-dotenv
google-generativeai
youtube-transcript-api
fpdf
requests
beautifulsoup4
PyMuPDF
transformers
pandas
numpy